* commit '23ae6dbb50b34e7630172bda5a2eba519f11c6ff':
  Integrate creation of Zip report archives.
  clean up
  restructured code and clean up
  removed commented out code
  added necessary imports for editing zip files
  commented out creation of installation report
  adjusted test report skript for zip upload to nexus
  added script to store zip file with raw data in nexus

+++ src/org/ods/util/FileUtil.groovy
@@ -1,5 +1,10 @@
 package org.ods.util
 
+@Grab(group="net.lingala.zip4j", module="zip4j", version="2.1.1")
+
+import net.lingala.zip4j.ZipFile
+import net.lingala.zip4j.model.ZipParameters
+
 class FileUtil {
     static File createTempFile(String baseDir, String prefix, String suffix, byte[] data) {
         def tmpFile = null
@@ -11,14 +16,30 @@ class FileUtil {
 
             // Create a temporary file containing data
             tmpFile = File.createTempFile("${prefix}-", "-${suffix}", dir) << data
-        } catch (Throwable e) {
-            throw e
-
-            if (tmpFile) {
+        } finally {
+            if (tmpFile && tmpFile.exists()) {
                 tmpFile.delete()
             }
         }
 
         return tmpFile
+    }
+
+    static byte[] createZipFile(String path, Map<String, byte[]> files) {
+        // Create parent directory if needed
+        def baseDir = new File(path).getParentFile()
+        if (!baseDir.exists()) {
+            baseDir.mkdirs()
+        }
+
+        // Create the Zip file
+        def zipFile = new ZipFile(path)
+        files.each { filePath, fileData ->
+            def params = new ZipParameters()
+            params.setFileNameInZip(filePath)
+            zipFile.addStream(new ByteArrayInputStream(fileData), params)
+        }
+
+        return new File(path).getBytes()
     }
 }
